Stub out all Viscara NPC abilities. Implement Earthquake, Fire Breath, and Iron Shell. Fix a number of bugs with timestamp time zones and positioning of abilities. Add tracking for the type of Busy status.

+++ SWLOR.Game.Server/Feature/AbilityDefinition/Force/ForceLightningAbilityDefinition.cs
@@ -54,7 +54,7 @@ namespace SWLOR.Game.Server.Feature.AbilityDefinition.Force
                 ApplyEffectToObject(DurationType.Instant, EffectDamage(damage), target);
                 ApplyEffectToObject(DurationType.Instant, EffectVisualEffect(VisualEffect.Vfx_Beam_Lightning), target);
             });
-            ActionCastFakeSpellAtObject((int)Spell.LightningBolt, target);
+            ActionCastFakeSpellAtObject(Spell.LightningBolt, target);
             
             Enmity.ModifyEnmityOnAll(activator, 1);
             CombatPoint.AddCombatPointToAllTagged(activator, SkillType.Force, 3);
